restructure files in graphql folder
