+++ gridsome/lib/app/PluginStore.js
@@ -7,7 +7,7 @@ const camelCase = require('camelcase')
 const pathToRegexp = require('path-to-regexp')
 const slugify = require('@sindresorhus/slugify')
 const { NODE_FIELDS } = require('../utils/constants')
-const { parsePageQuery } = require('../graphql/utils/page-query')
+const { parsePageQuery } = require('../graphql/pageQuery')
 const { mapValues, cloneDeep } = require('lodash')
 const { cache, nodeCache } = require('../utils/cache')
 const { log, warn } = require('../utils/log')
