+++ gridsome/lib/graphql/interfaces.js
