+++ gridsome/lib/graphql/resolvers.js
